requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
google-generativeai==0.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
pyaudio==0.2.13
gTTS==2.3.2
pygame==2.5.0
googletrans==4.0.0-rc1
SpeechRecognition==3.10.0
python-dotenv==1.0.0
